Minor optimization: faster plink rsids QC.
We check whether plink data were annotated with rsids (if matching between the files is done on rsids) just as sanity check. If data were _not_ annotated with rsids, rsid column will be filled with ".". Full correct check takes 15-30 seconds, even though this is a minor QC, and we can be 100% sure that annotation exists just by looking at the first few lines of the plink file. So, this microoptimisation was born.